+++ src/cpp/galay-kernel/parallel/parallel_scheduler.h
@@ -25,7 +25,7 @@
 #include <cstddef>
 #include <thread>
 #include <atomic>
-#include <concurrentqueue/moodycamel/blockingconcurrentqueue.h>
+#include <galay/thirdparty/concurrentqueue/moodycamel/blockingconcurrentqueue.h>
 
 namespace galay::kernel
 {