+++ src/cpp/galay-mysql/async/conn_pool.h
@@ -16,7 +16,7 @@
 #include "../../galay-kernel/core/io_scheduler.hpp"
 #include "../../galay-kernel/core/task.h"
 #include "../../galay-kernel/core/waker.h"
-#include <concurrentqueue/moodycamel/concurrentqueue.h>
+#include <galay/thirdparty/concurrentqueue/moodycamel/concurrentqueue.h>
 #include <memory>
 #include <vector>
 #include <atomic>